:art: update code

+++ handyrec/features/type.py
@@ -31,10 +31,10 @@ class DenseFeature(Feature):
     name : str
         Name of feature, each feature should have a distinct name.
     dtype : str, optional
-        Data type, by default ``"int32"``.
+        Data type, by default ``"float32"``.
     """
 
-    def __init__(self, name: str, dim: int = 1, dtype: str = "int32"):
+    def __init__(self, name: str, dim: int = 1, dtype: str = "float32"):
         """Initialize a `DenseFeature`.
 
         Parameters
@@ -44,7 +44,7 @@ class DenseFeature(Feature):
         dim : int, optional
             Dimension of feature, by default ``1``.
         dtype : str, optional
-            Data type, by default ``"int32"``.
+            Data type, by default ``"float32"``.
         """
         super().__init__(name, dtype)
         self.dim = dim
